Patch to address column definitions.
In the v4.2.4 schema the following columns are listed as 'text' type columns.

.   learning_outcome_dim.description
.   module_dim.name

In the MySQL specific liquid files, the learning_outcome_dim table was failing
to be generated at all (due to a DDL syntax error).  The module_dim.name
column was too short for some of our actual data and that was generating an
error.  I went back to the `schema.json` and made adjustment to both file
(fixing the syntax error and adjusting the column spec).

I also applied similar changes to the MSSQL, Oracle, and PostgresQL variants.
I am unable to verify that these are syntactically correct, but they match
the similar constructs for other tables in their respective dialects.